% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.R
\name{colormap}
\alias{colormap}
\title{Plot a pseudo 3-dimensional Plot with the mean z Channel Coordinate as fill Color}
\usage{
colormap(template, sample, z = NULL, limits = NULL)
}
\arguments{
\item{template}{A 3-dimensional FSFTemplate}

\item{sample}{The sample to plot}

\item{z}{The channel that should be used as fill color}
}
\value{
A ggplot2 plot
}
\description{
Plot a pseudo 3-dimensional Plot with the mean z Channel Coordinate as fill Color
}
