% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.R
\name{colormapFF}
\alias{colormapFF}
\title{Plot A flowCore flowFrame (3 channels)}
\usage{
colormapFF(
  flowFrame,
  channels,
  z = NULL,
  resolution = 50,
  transformation = log10,
  ...
)
}
\arguments{
\item{flowFrame}{A \code{\link{flowCore}} flowFrame}

\item{channels}{A vector of length 3 containing the channel names}

\item{z}{The channel that should be used as fill color}

\item{resolution}{Resolution of the plot}

\item{transformation}{Transformation. Default is log10. transformation = NULL for no transformation}

\item{...}{Additional arguments passed to plot.FSFTemplate}
}
\value{
A ggplot2 plot
}
\description{
Plot A flowCore flowFrame (3 channels)
}
