% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreTemplate.R
\docType{class}
\name{CoreTemplate-class}
\alias{CoreTemplate-class}
\title{The S4 CoreTemplate class}
\value{

}
\description{
The S4 CoreTemplate class
}
\section{Slots}{

\describe{
\item{\code{coords}}{A data.frame containing the coordinates of every bin}

\item{\code{counts}}{A data.frame containing the number of events for every bin}
}}

