% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreTemplate.R
\name{CoreTemplate}
\alias{CoreTemplate}
\title{Create A CoreTemplate Object}
\usage{
CoreTemplate(flowset, channels, resolution = 4)
}
\arguments{
\item{flowset}{A flowCore \code{flowSet}}

\item{channels}{A vector of channel names}

\item{resolution}{A number specifying the number of bins on one axis}
}
\value{
A CoreTemplate
}
\description{
Create A CoreTemplate Object
}
