% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Convenience.R
\name{frequencies}
\alias{frequencies}
\title{Get Table of Frequencies of every Bin's Event for Every Sample}
\usage{
frequencies(template)
}
\arguments{
\item{template}{A \code{\link{CoreTemplate}}}
}
\value{
a data.frame
}
\description{
Get Table of Frequencies of every Bin's Event for Every Sample
}
