% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FSFTemplate.R
\docType{class}
\name{FSFTemplate-class}
\alias{FSFTemplate-class}
\title{The S4 FSFTemplate class}
\value{

}
\description{
The S4 FSFTemplate class
}
\section{Slots}{

\describe{
\item{\code{coords}}{A data.frame containing the coordinates of every bin}

\item{\code{counts}}{A data.frame containing the number of events for every bin}
}}

