% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FSFTemplate.R
\name{FSFTemplate}
\alias{FSFTemplate}
\title{Create A FSFTemplate Object}
\usage{
FSFTemplate(
  flowset,
  channels,
  resolution = 4,
  transformation = log10,
  verbose = T
)
}
\arguments{
\item{flowset}{A flowCore \code{flowSet}}

\item{channels}{A vector of channel names}

\item{resolution}{A number specifying the number of bins on one axis}
}
\value{
A FSFTemplate
}
\description{
Create A FSFTemplate Object
}
