% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Convenience.R
\name{getCT}
\alias{getCT}
\title{Extract a CoreTemplate from a NDTemplate based on Channel Names}
\usage{
getCT(template, channels)
}
\arguments{
\item{template}{A \code{\link{NDTemplate}}}

\item{channels}{A vector of channel names}
}
\value{
A \code{\link{CoreTemplate}}
}
\description{
Extract a CoreTemplate from a NDTemplate based on Channel Names
}
