% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NDTemplate.R
\docType{class}
\name{NDTemplate-class}
\alias{NDTemplate-class}
\title{The S4 NDTemplate class}
\value{

}
\description{
The S4 NDTemplate class
}
\section{Slots}{

\describe{
\item{\code{templates}}{A list of combinations of \code{CoreTemplate}s. Sorted by number of dimensions.}

\item{\code{metadata}}{A metadata table}

\item{\code{resolution}}{The number of bins on one axis}

\item{\code{nSamples}}{The number of samples}
}}

\examples{

}
