% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NDTemplate.R
\name{NDTemplate}
\alias{NDTemplate}
\title{Create A NDTemplate Object}
\usage{
NDTemplate(flowset, channels, resolution = 4, metadata = data.frame())
}
\arguments{
\item{flowset}{A flowCore \code{flowSet}}

\item{channels}{A vector of channel names}

\item{resolution}{A number specifying the number of bins on one axis}

\item{metadata}{A metadata table}
}
\value{
A NDTemplate
}
\description{
Create A NDTemplate Object
}
\examples{

}
