% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.R
\name{plot.CoreTemplate}
\alias{plot.CoreTemplate}
\title{Plot A 1- to 3-Dimensional CoreTemplate}
\usage{
\method{plot}{CoreTemplate}(template, sample)
}
\arguments{
\item{template}{The \code{\link{CoreTemplate}} to plot}

\item{sample}{The sample to plot}
}
\value{
Depending on the number of dimensions nothing, a ggplot2 object or a plotly object
}
\description{
Plot A 1- to 3-Dimensional CoreTemplate
}
