% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.R
\name{plot.NDTemplate}
\alias{plot.NDTemplate}
\title{Plot Every 1- Or 2-Dimensional Combination Of Axes Of A NDTemplate}
\usage{
\method{plot}{NDTemplate}(template, sample, dimen = 1)
}
\arguments{
\item{template}{The NDTemplate to plot}

\item{sample}{The sample to plot}

\item{dimen}{The number of axes to plot (1 or 2)}
}
\value{

}
\description{
Plot Every 1- Or 2-Dimensional Combination Of Axes Of A NDTemplate
}
