% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NSR.R
\name{plot_nsr}
\alias{plot_nsr}
\title{Graphical representation of the calculated MNSRs over the respective grid resolutions. Highlighting of the optimal MNSR value.}
\usage{
plot_nsr(obj, ylimit = NULL, xtext = NULL, ytext = NULL, legend = "topright")
}
\arguments{
\item{obj}{The output of \code{\link{vario_nsr}}}

\item{ylimit}{Numerical value. y-axis limit}

\item{xtext}{Numerical value. x-coordinate for the text within the graphic, which marks the optimal MNSR value. Default is NULL, so the value is automatically set depending on the position of the optimal MNSR_value.}

\item{ytext}{Numerical value. Analogous to xtext}

\item{legend}{Position of the legend. See \code{\link{legend}}. Default is "topright".}
}
\value{
A graph showing the MNSR values of \code{\link{vario_nsr}} depending on the respective grid resolution. In addition, the optimal MNSR value is shown on the corresponding grid resolution with a red line with the lettering which indicates the respective grid resolution.
}
\description{
Graphical representation of the calculated MNSRs over the respective grid resolutions. Highlighting of the optimal MNSR value.
}
