% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.R
\name{plotFF}
\alias{plotFF}
\title{Plot A flowCore flowFrame}
\usage{
plotFF(flowFrame, x, y, resolution = 50, transformation = log10, ...)
}
\arguments{
\item{flowFrame}{A \code{\link{flowCore}} flowFrame}

\item{x}{x-channel}

\item{y}{Y-channel}

\item{resolution}{Resolution of the plot}

\item{transformation}{Transformation. Default is log10. transformation = NULL for no transformation}

\item{...}{Additional arguments passed to plot.FSFTemplate}
}
\value{
A ggplot2 plot
}
\description{
Plot A flowCore flowFrame
}
