% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.R
\name{plotTscores}
\alias{plotTscores}
\title{Plot t-scores in 1 or 2 Dimensions}
\usage{
plotTscores(template, ts, limits = NULL)
}
\arguments{
\item{template}{A FSFTemplate Object}

\item{ts}{A vector of t-scores}

\item{limits}{Limits of the color scale}
}
\value{
A ggplot Object
}
\description{
Plot t-scores in 1 or 2 Dimensions
}
