% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Convenience.R
\name{print.FSFTemplate}
\alias{print.FSFTemplate}
\title{Print Info about a FSFTemplate}
\usage{
\method{print}{FSFTemplate}(template, samples = 1)
}
\arguments{
\item{template}{A FSFTemplate}
}
\value{

}
\description{
Print Info about a FSFTemplate
}
