% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Convenience.R
\name{pw.adonis2}
\alias{pw.adonis2}
\title{Simplified pairwise adonis2 for only one Term}
\usage{
pw.adonis2(distM, term, data, adjust = "bonferroni")
}
\arguments{
\item{distM}{A distance Matrix}

\item{term}{The term to compute \code{\link{adonis2}} for}

\item{data}{The data from which to take the term}

\item{adjust}{p-value adjustment method for \code{\link{p.adjust}}}
}
\value{
Printed Output
}
\description{
Simplified pairwise adonis2 for only one Term
}
