% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Convenience.R
\name{shrink}
\alias{shrink}
\title{Create lower Dimension FSFTemplate from higher Dimension FSFTemplate through Addition.}
\usage{
shrink(template, channels)
}
\arguments{
\item{template}{A FSFTemplate}

\item{channels}{Channels of the new FSFTemplate}
}
\value{
A FSFTemplate
}
\description{
Create lower Dimension FSFTemplate from higher Dimension FSFTemplate through Addition.
}
