% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Convenience.R
\name{significantChannels}
\alias{significantChannels}
\title{Check which flow cytometry channels show significant difference between given metadata groups}
\usage{
significantChannels(template, term, data, adjust = "bonferroni", ...)
}
\arguments{
\item{template}{A FSFTemplate}

\item{term}{The term to compute \code{\link{adonis2}} for}

\item{data}{The data from which to take the term}

\item{adjust}{p-value adjustment method for \code{\link{p.adjust}}}

\item{...}{Additional arguments passed to \code{\link{adonis2}}}
}
\value{
A ggplot2 object
}
\description{
Check which flow cytometry channels show significant difference between given metadata groups
}
