% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NSR.R
\name{vario_func}
\alias{vario_func}
\title{Calculates semivariograms for a grid resolution (by FSFTemplate) for all available individuals and models them successively by the given models}
\usage{
vario_func(modells, template, min = 20, endCut = NULL, cores = 1)
}
\arguments{
\item{modells}{Character vector that can include the following models: "gaussian", "exponential", "cauchy", "circular", "cubic", "matern", "spherical", "linear", "gneiting", "wave".}

\item{template}{FSFTemplate object}

\item{min}{Integer. Minimum number of events in an interval. If an interval does not meet the minimum number, it is merged with the previous interval. If the first interval does not meet the minimum number, it will be merged with the next interval that meets the minimum number.}

\item{endCut}{Numeric value. A distance specification from which values are forced into an interval in the semivariogram.}

\item{cores}{Integer. Threads to use.}
}
\value{
List containing \enumerate{
\item ergValues - Dataframe with best model, nugget, sill, ML value and NSR value.
\item MNSR - Mean NSR-Value
\item MML - Mean maximum likelihood value
}
}
\description{
The NSR = nugget-sill-ratio. The NSR takes values in the interval [0,1].
The smaller the value, the smaller is the loss of information due to measurement errors, or due to the accumulation of data in range units (e.g. by a grid).
The smaller the NSR, the greater the spatial dependence of the data. The MNSR is the mean value of the NSRs generated for each individual.
Note: The NSR depends, among other things, on the grid structure, the generated distance intervals in the semivariogram and the used models!
}
