% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NSR.R
\name{vario_nsr}
\alias{vario_nsr}
\title{Applies vario_func for an arbitrary number of grid resolutions}
\usage{
vario_nsr(
  flowset,
  channels,
  resolutions,
  transformation = log10,
  verbose = TRUE,
  modells,
  min = 20,
  endCut = NULL,
  cores = 1
)
}
\arguments{
\item{flowset}{A \code{\link{flowSet}}}

\item{channels}{Vector of flow cytometry channels}

\item{resolutions}{numerical vector of grid resolutions to check}

\item{transformation}{Transformation function passed to \code{\link{FSFTemplate}}}

\item{verbose}{Boolean. Passed to \code{\link{FSFTemplate}}}

\item{modells}{Character vector passed to \code{\link{vario_func}}}

\item{min}{Integer. Minimum number of events in an interval. Passed to \code{\link{vario_func}}}

\item{endCut}{Numerical value. Passed to \code{\link{vario_func}}}

\item{cores}{Integer. Threads to use.}
}
\value{
Data.frame containing \enumerate{
\item resolution - FSFTemplate grid resolution
\item MNSR - Mean NSR-Value
\item MML - Mean maximum likelihood value
}
}
\description{
Applies vario_func for an arbitrary number of grid resolutions
}
