% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WeightedBray.R
\name{weightedBray}
\alias{weightedBray}
\title{Compute Weighted Bray Distance between Samples of a CoreTemplate}
\usage{
weightedBray(template, gamma = 8, w = NA)
}
\arguments{
\item{template}{A \code{\link{CoreTemplate}} Object}

\item{gamma}{A positive number specifying the drop off of the weight matrix when w = NA}

\item{w}{Optionally you may provide your own weight matrix created with \code{\link{weightMatrix}}}
}
\value{
A distance matrix
}
\description{
Compute Weighted Bray Distance between Samples of a CoreTemplate
}
