% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tscores.R
\name{tscores}
\alias{tscores}
\title{Get t-scores according to Welch's t-statistic for a FSFTemplate and a Metadata Group}
\usage{
tscores(template, group)
}
\arguments{
\item{template}{A FSFTemplate Object}

\item{group}{A vector of metadata}
}
\value{
A DataFrame of t-scores
}
\description{
Get t-scores according to Welch's t-statistic for a FSFTemplate and a Metadata Group
}
